pyproject.toml
[project]
name = "etcd3aio"
version = "0.1.0"
description = "Async Python client for etcd v3 (grpc.aio, Python 3.13+)"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
keywords = ["etcd", "etcd3", "asyncio", "grpc", "distributed"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: System :: Distributed Computing",
    "Framework :: AsyncIO",
]
dependencies = [
    "grpcio>=1.78.1",
    "protobuf>=6.33.5",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
    "pyright",
    "pyinstaller",
    "grpcio-tools",
    "pytest-asyncio",
]

[project.scripts]
#etcd3aio = ""

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]
exclude = ["build", "dist", "__pycache__", "*.egg-info", ".venv", "src/etcd3aio/proto"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "PL", "PT"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["PLR2004"]

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
include = ["src", "tests"]
extraPaths = ["src"]
exclude = ["build", "dist", "**/__pycache__", "**/*.egg-info", ".venv", "src/etcd3aio/proto"]
venvPath = "."
venv = ".venv"
reportMissingImports = "none"
reportMissingModuleSource = "none"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=xml --cov-report=term --cov-report=html -v"

[tool.coverage.run]
omit = [
    "src/etcd3aio/proto/*",
]
